Use is_valid_address() in place of is_address and move it to util.py
